+++ main.go
@@ -66,8 +66,9 @@ func (a *application) createHTTPServer(
 		router.Path("/healthz").Handler(libhttp.NewPrintHandler("OK"))
 		router.Path("/readiness").Handler(libhttp.NewPrintHandler("OK"))
 		router.Path("/metrics").Handler(promhttp.Handler())
-		router.Path("/setloglevel/{level}").
-			Handler(log.NewSetLoglevelHandler(ctx, log.NewLogLevelSetter(2, 5*time.Minute)))
+		router.Path("/setloglevel/{level}").Handler(
+			log.NewSetLoglevelHandler(ctx, log.NewLogLevelSetter(2, 5*time.Minute)),
+		)
 		router.Path("/read").Handler(factory.CreateReadHandler(sentryClient, saramaClient))
 
 		glog.V(2).Infof("starting http server listen on %s", a.Listen)