+++ pkg/handler.go
@@ -34,7 +34,6 @@ func NewHandler(
 			if topic == "" {
 				return errors.New(ctx, "parameter topic missing")
 			}
-
 			offset, err := libkafka.ParseOffset(ctx, req.FormValue("offset"))
 			if err != nil {
 				return errors.Wrap(ctx, err, "parse parameter offset failed")
@@ -47,15 +46,16 @@ func NewHandler(
 			if err != nil {
 				return errors.Wrap(ctx, err, "parse parameter partition failed")
 			}
-
 			filterValue := req.FormValue("filter")
 			if len(filterValue) > 1024 {
 				return errors.New(ctx, "filter parameter exceeds maximum length of 1024 bytes")
 			}
 			filter := []byte(filterValue)
 
-			glog.V(2).
-				Infof("read records from topic %s and partition %d and offset %d with limit %d started", topic, partition.Int32(), offset.Int64(), limit)
+			glog.V(2).Infof(
+				"read records from topic %s and partition %d and offset %d with limit %d started",
+				topic, partition.Int32(), offset.Int64(), limit,
+			)
 			changes, err := changesProvider.Changes(ctx, topic, *partition, *offset, limit, filter)
 			if err != nil {
 				if !errors.Is(err, sarama.ErrOffsetOutOfRange) {
@@ -82,11 +82,13 @@ func NewHandler(
 				Records:    changes,
 				NextOffset: &nextOffset,
 			}
-			if err := SendJSONResponse(resp, page, http.StatusOK); err != nil {
+			if err := libhttp.SendJSONResponse(ctx, resp, page, http.StatusOK); err != nil {
 				return errors.Wrap(ctx, err, "send json failed")
 			}
-			glog.V(2).
-				Infof("read %d records from topic %s and partition %d and offset %d with limit %d completed", len(page.Records), topic, partition.Int32(), offset.Int64(), limit)
+			glog.V(2).Infof(
+				"read %d records from topic %s and partition %d and offset %d with limit %d completed",
+				len(page.Records), topic, partition.Int32(), offset.Int64(), limit,
+			)
 			return nil
 		},
 	)